feat: replace Clapper with mpv (libmpv) video renderer
Add an mpv-based renderer as the new default video backend and drop the
Clapper widget. A `video-renderer` GSettings key selects between mpv and
the GStreamer gtk4paintablesink fallback; switching it rebuilds the
active wallpaper live. mpv is gated behind the optional `mpv` cargo
feature / `-Dmpv` meson option so builds without libmpv fall back to
GStreamer.

Also bump gtk/gst/webkit deps, switch the renderer volume API to a 0-100
int, and reduce content-fit to Fill/Contain/Cover.

+++ src/widget/web.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2026  Jeff Shee
+// Copyright (C) 2026 Jeff Shee <jeffshee8969@gmail.com>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -63,7 +63,7 @@ impl RendererWidget for WebWidget {
 
     fn stop(&self) {}
 
-    fn set_volume(&self, _volume: f64) {}
+    fn set_volume(&self, _volume: i32) {}
 
     fn set_mute(&self, _mute: bool) {}
